consumers/mongodb/requirements.txt
pymongo==4.0.1
kafka-python==2.0.2